# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Ocytosign",
        "description": "Documentation de l'API Ocytosign",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/interpretres": {
            "get": {
                "tags": [
                    "Interprètes"
                ],
                "summary": "Liste des interprètes",
                "operationId": "5d8b7285074a060c35b1f5fb5ddb81b1",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Interprètes"
                ],
                "summary": "Créer un interprète",
                "operationId": "2f39bd903aabe80f735dd1497bb1c374",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_medecin",
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "id_medecin": {
                                        "type": "integer"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interprète créé avec succès"
                    }
                }
            }
        },
        "/api/interpretres/{id}": {
            "get": {
                "tags": [
                    "Interprètes"
                ],
                "summary": "Afficher un interprète",
                "operationId": "919edbae12824f8d893edaf2e3a46f0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interprète trouvé"
                    },
                    "404": {
                        "description": "Interprète non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Interprètes"
                ],
                "summary": "Modifier un interprète",
                "operationId": "b8490b48b01a62d2ecbf28ed5eb698e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_medecin",
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "id_medecin": {
                                        "type": "integer"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interprète mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Interprètes"
                ],
                "summary": "Supprimer un interprète",
                "operationId": "1f387bc8a815558ad5544c4beae0e06d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interprète supprimé"
                    },
                    "404": {
                        "description": "Interprète non trouvé"
                    }
                }
            }
        },
        "/api/medecins": {
            "get": {
                "tags": [
                    "Médecins"
                ],
                "summary": "Liste des médecins",
                "operationId": "6ab602fc805504c7805644cf370f89cd",
                "responses": {
                    "200": {
                        "description": "Liste des médecins récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Médecins"
                ],
                "summary": "Créer un nouveau médecin",
                "operationId": "999425c412ff82050637d174c566a2db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Médecin créé avec succès"
                    }
                }
            }
        },
        "/api/medecins/{id}": {
            "get": {
                "tags": [
                    "Médecins"
                ],
                "summary": "Afficher un médecin",
                "operationId": "8667128b9a43942ac6386ee1c0befde5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Médecin trouvé"
                    },
                    "404": {
                        "description": "Médecin non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Médecins"
                ],
                "summary": "Mettre à jour un médecin",
                "operationId": "59c0f78c0afeab512befc9a84ad71a61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Médecin mis à jour avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Médecins"
                ],
                "summary": "Supprimer un médecin",
                "operationId": "842cacfbf68c23c09eff54e6c1949e8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Médecin supprimé avec succès"
                    },
                    "404": {
                        "description": "Médecin non trouvé"
                    }
                }
            }
        },
...
    },
...
}